chore: fix NodeJS init templates V2 unit tests for new-style synthesis (#14400)
Our NodeJS init templates for 'app' and 'lib' come with a unit test that verifies the synthesized Stack is empty.
However, when we switched to new-style synthesis by default in V2 in #14182,
that assertion is no longer true -
new-style synthesis always adds a Parameter and a Rule to the template.

Change the assertion to verify that the `Resources` section is empty instead.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*

(cherry picked from commit a1072aef8ff7a4a5e7380b7a20d76456f50b640f)