chore: strip removed experimental modules from CHANGELOG entries
For the developer preview release of CDKv2, we are stripping experimental
modules from the packaged release.  However, the entries from the CHANGELOG are
still included, as `standard-release` has no knowledge of this process.  This
script -- run as part of the 'bump' process -- attempts to identify and remove
CHANGELOG lines that are associated with experimental modules.

This is far from perfect, but might work well enough for our developer preview
releases, until we have finalized our experimental modules release
plan. Feedback very welcome.