chore: fix failing v2 Java init template test
Similar to #14400, which fixed the failing JavaScript and TypeScript init
template tests. The test fails due to the new Parameter and Rule inserted by the
new-style synthesis, which means the stack is no longer empty.

Updated the test to check that there are no resources present by default, which
frankly feels a bit odd, but at least is consistent.

Separately, we need to dive deeper into the fact that `mvn package` can fail
when installing the Java template, but `cdk init` still succeeds.