chore: make github-helpers major-version aware (#14422)
The github-helpers fetchPreviousVersion function is used during the regression
tests to determine the previous published version of the CDK to test against.

This script was previously not major-version aware, meaning that the regressions
could be comparing v1 to v2, leading to errors like:
"No matching version found for @aws-cdk/aws-cloudformation@2.0.0-rc.1".
(using v1 packages with v2 versioning).

This changes the script to be major-version aware, and attempt to return the
most recent version that matches the same major version, falling back to the
most recent release if no major-version matches are found.

_Note_ - Will backport this to master.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*