feat(secretsmanager): Automatically grant permissions to rotation Lambda
When you use the AWS Secrets Manager console to configure rotation for a
secret for one of the fully supported databases, the console configures
almost all parameters for you. But if you create a function or opt to do
anything manually for other reasons, you also might have to manually
configure the permissions for that part of the rotation.

https://docs.aws.amazon.com/secretsmanager/latest/userguide/rotating-secrets-required-permissions.html