fix(cfn-include): correctly parse Fn::Sub expressions containing serialized JSON
Our logic for parsing `${}` expressions inside `Fn::Sub` incorrectly looked for closing braces everywhere in the string,
instead of doing it to the right of the opening `${`.
This would fail if the string passed to `Fn::Sub` was serialized JSON
(as those would contain closing braces completely unrelated to the `${`).

Fixes #14095