chore: remove symlink from version control
The combination of CodePipeline+CodeBuild will not properly extract
symlinks from a git repository. Instead, the symlink will be replaced
with a textual representation of its target.

To properly test the symlink behavior, we need to remove the symlink
from version control and re-create it at test time, after
CodePipeline+CodeBuild have extracted the archive.