chore: remove symlink from version control (#14520)
The combination of CodePipeline+CodeBuild will not properly extract
symlinks from a git repository. Instead, the symlink will be replaced
with a textual representation of its target.

To properly test the symlink behavior, we need to remove the symlink
from version control and re-create it at test time, after
CodePipeline+CodeBuild have extracted the archive.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*