fix(cfn-include): correctly parse Fn::Sub expressions containing serialized JSON (#14512)
Our logic for parsing `${}` expressions inside `Fn::Sub` incorrectly looked for closing braces everywhere in the string,
instead of doing it to the right of the opening `${`.
This would fail if the string passed to `Fn::Sub` was serialized JSON
(as those would contain closing braces completely unrelated to the `${`).

Fixes #14095

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*