fix(apigatewayv2): incorrect metric names for client and server-side errors
In the HTTP API offering of API Gateway, the [metric names][1] for
client-side and server-side errors are '4xx' and '5xx'.

This was incorrectly configured as '4XXError' and '5XXError' which is
the [metric name in the REST API offering][2].

fixes #14503

In addition, a previous change elevated metrics defined in HTTP API such
that it is also available for the Websocket API offering. Unfortunately,
the set of metrics or metric names available in the [HTTP API][1] and
[Websocket API][3] offerings do not match up.

Hence, remove these metric names from Websocket APIs.

[1]: https://docs.aws.amazon.com/apigateway/latest/developerguide/http-api-metrics.html
[2]: https://docs.aws.amazon.com/apigateway/latest/developerguide/api-gateway-metrics-and-dimensions.html
[3]: https://docs.aws.amazon.com/apigateway/latest/developerguide/apigateway-websocket-api-logging.html

BREAKING CHANGES: The `metricXXX` methods are no longer available in the
`IApi` interface. The existing ones are moved into `IHttpApi` and new
ones will be added to `IWebsocketApi`.
* **apigatewayv2:** The `metricXXX` methods are no longer available in
the `IStage` interface. The existing ones are moved into `IHttpStage`
and new ones will be added to the `IWebsocketStage`.