chore(cloudfront): remove the use of calculateFunctionHash
`calculateFunctionHash()` was used to compute the 'refresh token' of the
custom resource for the EdgeFunction construct.

This method is private to the lambda module and is deemed to be changed.
Instead, use the lambda function version's logical id.

The logical id of the version is computed using the same mechanism and
is a more reliable determinant of whether the underlying function
version changed.