chore(custom-resources): import the AWSLambda package explicitly (#14643)
When linking the aws-cdk repository to a CDK app using the `link-all.sh` script,
if the app uses `ts-node`,
the Lambda code in the @aws-cdk/custom-resources package gets picked up by the TypeScript compiler.
That code relied on the `aws-lambda` package being implicitly available,
but that would cause `ts-node` to fail.
Add an explicit import of it in the code -
I checked the only difference in the generated JS code is the sourceMappingUrl,
so it shouldn't make a difference at runtime,
but allows `ts-node` to load that file successfully.

Fixes #11627

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*