chore(msk): add ignore-assets pragma to cluster integ test (#14725)
The MSK module relies on custom resources, which in turn create a Lambda
function with assets. The way the current (lerna/yarn) build works includes
the .ts file (as well as the .d.ts and .js) files in the asset bundle. Using the
new `nozem` build (correctly) only includes the .d.ts and .js files, leading to a
different asset hash.

Since we don't care about the actual hash anyway, adding the ignore-assets
pragma so this test can pass with either build tool.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*