fix(core): cannot determine packaging when bundling that produces an archive is skipped (#14372)
Bundling is skipped when the bundling directory already exists on disk
(cache). If it previously produced a single archive file that has been
moved to the staging directory the current logic either failed to
determine the correct packaging type or threw because it expected the
bundling directory to include a single archive file.

Fix it by "touching" the archive file in the bundling directory after it
has been moved to the staging directory.

Fixes #14369


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*