fix(pipelines): synth fails if 'aws-cdk' is not in `package.json`
Our `aws-sdk` dependency has recently added a `postinstall` script to
their package.

For whatever reason, if the `unsafe_perm` NPM config flag is not set to
`true` and `npx` needs to install the `aws-cdk` package (because it's
not in the user's `package.json` and so didn't get installed as part of
`npm ci`), then the `postinstall` script fails.

I've confirmed that this fixes it but I haven't 100% put my finger on
the smoking gun yet.

* `npx package-without-postinstall` works fine, even without this
  setting.
* `npm install -g package-with-postinstall` also works fine, even though
  it's still installed as `root`.

It seems to be the combination of `npx`, `root` and `postinstall` that
is causing the issue.

Fixes #14658.