chore(docs): added troubleshooting entry for Docker error in the synthesis step (#14793)
Documents the action the user should take to prevent #9217.

The problem here is that, for any app that does asset bundling or uses constructs like `NodeJsFunction` (or its counterparts in other languages), the CodeBuild project has to be configured to run in privileged mode. But the only way to check whether the app uses these things is to run it and check the generated cloud assembly, which will fail if the privileged flag is not set.

The only way to break out of this circular dependency would be to perform some sort of static analysis in the app before running it, or perhaps creating some concept of dry-run, in which the app just reports what it would do without actually doing it. But these solutions would be very complex to implement and probably not worth it. So we're resorting to ask the user to tell the CDK what to do.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*