chore(cdk): increase timeout for all bootstrapping integ tests
In a previous commit - 8395d9d1 - the timeout for a single test was
increased to 1 hour since it was noticed to fail consistently.

Since then, other tests have also failed intermittently due to
CloudFormation actions taking longer than expected in some regions.

Increase the timeout for all tests to 1 hour when executing in
CodeBuild. To keep the dev-test cycle reasonable, maintain a 10 minute
timeout when developing outside of CodeBuild.