chore: remove prlinter module and merge it with prlint (#14825)
Previously, the actual linter is modeled as a regular npm package named
'prlint' and a shim module named 'prlinter' existed that exposed this as
a Github action package.

This extra layer of indirection is not very useful until we have another
use case where the linter will be used outside of Github Actions.

For the forseeable future, this will be only used via Github Actions.
Remove this extra layer to simplify.