fix(rds): Add exception throw when az is defined for multi-az db instance (#14837)
This PR is intended to cover a small bug that is produced when you create a RDS instance with both Multi-AZ and Availability Zone properties defined, which are incompatible.

That can cause RDS replace if Multi-AZ is setup after to an existing stack with AZ defined previously, because CDK removes AZ setup without any notice to the developer.

This contribution will give the developer visibility of this incompatibility so is not possible to apply both parameters at the same time.

More details in issue [#10949](https://github.com/aws/aws-cdk/issues/10949)

Closes [#10949](https://github.com/aws/aws-cdk/issues/10949)
----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*