fix(secretsmanager): remove condition key for lambda rotation functionArn
The policy needs to restrict which secrets the lambda function is
allowed to access/rotate. The official documentation recommends usage of
the `secretsmanager:resource/AllowRotationLambdaArn` context key and any
secret (`"Resource": "*"`):

> Filters the request based on the ARN of the Lambda rotation function
attached to the target resource of the request. This enables you to
restrict access to only those secrets with a rotation Lambda ARN
matching this value. Secrets without rotation enabled or with a
different rotation Lambda ARN do not match.

https://docs.aws.amazon.com/secretsmanager/latest/userguide/rotating-secrets-required-permissions.html#rotating-secrets-required-permissions-user-vs-function
https://docs.aws.amazon.com/secretsmanager/latest/userguide/reference_iam-permissions.html#iam-contextkeys

But CDK automatically adds a `DependsOn` on the role and default policy
to prevent race condition when creating policy and lambda in parallel
which caused a circular dependency https://github.com/aws/aws-cdk/issues/14868.

By removing the condition the circular dependency is gone but since the
policy declares the target secret as well it was anyway redundant. Other
than the documentation above, which only used the condition but allowed
rotation for all secrets (`resources: "*"`).