fix: http api default authorizer
The default authorizer worked by passing the authorizer config to routes in the api by the addRoutes method.

We completely forgot about the use case of the default integration, so currently using default integration + default authorizer does not create an authorizer.

This PR fixes the bug and allows using default authorizer + default integration as expected.