chore(prlint): ban 'breaking changes' clause on stable modules (#14861)
The `scripts/check-api-compatibility.sh` script prevents breaking API
changes in stable CDK modules.
However, this does not prevent functional breaking changes. It is
infeasible to actually detect and prevent this.

The CDK must be more strict on preventing such changes and the impact
due to their perception.

A linter error when a 'BREAKING CHANGE' clause for a stable module is
encountered in the PR is a good way to communicate this information.