fix(lambda-nodejs): cannot bundle locally when consuming a node module with a NodejsFunction
In #14739 changed how `esbuild` is run. It now uses the package manager.

When consuming a node module with a `NodejsFunction` we need to run
`esbuild` from the directory containing the lock file. This is where
we have `node_modules\.bin`. If we run it from the directory containing
the entry file the package manager doesn't "see" `node_modules\.bin`.