feat(pipelines): allow switching to one CodeBuild action for same-typed assets (#13803)
This PR refactors the Asset Stage of the CDK pipeline so that only one FileAsset and one DockerAsset Action are created that are uploading all assets of the respective type.

The main benefit of this is improved performance as uploading in parallel does not save enough time to compensate slower startups of multiple CodeBuilds.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*