fix(pipelines): self-update role assumes hard-coded role names
Right now, the selfupdate action assumes hardcoded `*-publishing-role-*` and
`*-deploy-role-*` roles, because we can't know the actual role names at the time
the policies are being created.

This means pipelines' self-update is unusable for customers with custom
bootstrap role names. To solve this -- and other -- issue(s), tag the roles to
identify them, and change out the policy to use these tags instead.

This requires bumping the bootstrap stack template version, and the minimum
required version in the default synthesizer.

fixes #14877
related #9271