(feat)(Service Catalog) Initial implementation of L2 constructs
This is the first experimental release of L2 constructs for service catalog.  Mainly "AWs::ServiceCatalog::Portfolio"
and associated constructs created via method calls, and to a lesser extent "AWS::ServiceCatalog::CloudFormationProduct"
which is renamed to a simpler "product" construct. We expect to iterate on these quickly with user feedback.

Updated README for pull request changes, minor refactoring.
Commit was squashed to properly add co-authors to actual commit and not just in github

Testing done
------------------
* `yarn build && yarn lint`
* `yarn test`
* `yarn integ`

Co-authored-by: Dillon Ponzo <dponzo18@gmail.com>
Co-authored-by: Jackie Wang <wanjacki@amazon.com>