fix(core): property overrides fail for references (#15018)
If a property override added an `IResolvable` as the value,
it would not be resolved,
but instead treated as a CloudFormation object when rendering the resource's properties,
which would lead to some weird behavior
(the properties of the `IResolvable` would be merged into the template).
This was noticed by a customer using CFN-Include.

As a solution, resolve the overrides before merging them with the properties of the resource.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*