fix(pipelines): assets buildspec can exceed 25k size limit (#14974)
When using the new singlePublisher mode for assets you can hit an undocumented limitation of CodeBuild. An inline buildspec has a size limit of 25600 characters.

To prevent this from happening this asset publishing mode now writes the buildspec to a file in the cloud assembly and references this file in the CodeBuild project.

Another benefit of this is that adding/removing assets does not force a pipeline mutation.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*