fix(cli): deployment error traceback overwritten by progress bar (#14812)
Currently, when the "stack activity monitor" is asked to stop, it does a final poll of the stack state and sends the new events to the progress-bar printer. However, the monitor already told the printer to stop, at which time it re-printed all failures. This means that the final poll will erroneously overwrite the failures (because the behaviour of the printer is to repeatedly print to the same console space). This change ensures the printer is asked to stop after the final poll occurs.

This change also removes the failure message for the root stack from the failure re-print, since it's not interesting information to surface to the user.

This change also re-prints failures for the legacy events-based printer.

fixes #14780

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*