feat(apigateway): disable execute api endpoint (#14526)
I noticed that [CloudFormation exposes the field DisableExecuteApiEndpoint on RestApi](https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-apigateway-restapi.html), but it isn't wired up in CDK (like it is for ApiGatewayv2). This would clean up our infrastructure code quite nicely, were you to accept it into main.

I've verified that CDK omits the field from the CloudFormation (CF) templates if it isn't defined and sets the specified value when it is. I also verified that when those CF templates are deployed to AWS, the default execute endpoint toggle is flipped in API Gateway.

There's also a unit test, I've verified that the fail condition works if the answer is incorrect.

Many thanks.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*