feat(servicecatalog): Initial implementation of L2 Portfolio
This is the first minimal release of an L2 construct for a service catalog portfolio. The portfolio
construct includes methods to give access to IAM Principals as well as share across accounts. In a future
PR, functionality to associate with a Product and add constraints will be added.

Testing done
------------------
* `yarn build && yarn lint`
* `yarn rosetta:extract --strict`
* `yarn test`
* `yarn integ`

Co-authored-by: Aidan Crank <arcrank@gmail.com>
Co-authored-by: Jackie Wang <wanjacki@amazon.com>