fix(cfn-include): NestedStack's Parameters are not converted to strings (#15098)
When we include a NestedStack when creating a CfnInclude instance,
the conversion of the Parameters property of the AWS::CloudFormation::Stack resource is performed manually
(because of the eventuality that one of those Parameters was also requested to be removed when including the nested stack).
In that manual conversion, we did not correctly convert the values to strings,
which is what the underlying CfnStack class expects.
And so, if the parent stack passed a non-string primitive value to a nested stack Parameter,
like a number or boolean, including the nested stack would fail with a validation exception.

Fixes #15092

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*