chore: make securityGroups a private property
Since the default security group is created within the events-targets
module, this means that unless a custom security group is passed in, the
public output property on the pattern in the default case would be
undefined. This change does not expose securtyGroups as a public output
property on the ScheduledTask pattern; they will be accessible via the
EcsTask. (See: https://github.com/aws/aws-cdk/pull/14953)