fix(core): Fix Duration rendering for hours as human string
When asked to convert full hours to human strings, due to floating point
errors, they get rendered as minutes. 1 hour converts to 60 minutes.
2 hours converts to 1 hour 60 minutes.

This request fixes that by first multiplying `amount` with the
`fromUnit.inMillis` before dividing by `toUnit.inMillis`.
This ensures that if the amount is big enough, it gets divided properly.