fix(ecs): TagParameterContainerImage cannot be used across accounts (#15073)
When TagParameterContainerImage is used from a different acocunt than the service itself is in,
it fails resolution, because the Task's execution Role does not have a physical name set by default.
Create it with PhysicalName.GENERATE_IF_NEEDED, so that it assigns a name when accessed from a cross-account context.

Fixes #15070

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*