feat(logs): make the addition of permissions to Lambda functions optional (#14222)
### What does this PR do?
Makes the addition of permissions to Lambda functions optional for users instantiating the LambdaDestination class. By default the permissions will be added in hopes of not breaking existing user's configurations. Added a unit test which tests the scenario in which a user would like to opt out of adding permissions to their function.

### Motivation
Fixes #14198 in which users with a high amount of lambdas will cause `The final policy size (20497) is bigger than the limit (20480)` in our [Datadog CDK Construct Library](https://github.com/DataDog/datadog-cdk-constructs)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*