feat(codestarnotifications): new L2 constructs (#10833)
Hi @skinny85 
Sorry for the late update PRs, and I am ready for the first run.

This PR includes:
`aws-codestarnotifications` For NotificationRule L2 construct.

`aws-codestarnotifications-targets` This library contains a set of classes which can be used as CodeStar notification targets and allows using `sns` or `chatbot` as notification rule targets(multiple targets).

`aws-codestarnotifications-source` This library contains a set of classes which can be used as CodeStar notification source and allows using `codebuild`, `codepipeline`, `codecommit` or `codedeploy` as a notification rule source(single source).

These are Separate packages for each target or each source can avoid the dependency cycle.

Resolves: #9680