chore: make the bump.js script fetch back only tags
The `bump.js` script removes all v1 tags locally before running `standard-version`,
and then brings them back.
The problem is that the bringing-back part was done with a regular `git fetch`,
which also fetches all remote branches from origin.
This is problematic for people, like me,
who don't want to see all of the janky personal branches that are in origin,
but are only interested in master/release/v2-main.

Change the `git fetch` command used in `bump.js` to only fetch tags.