feat(servicecatalog): initial implementation of the Portfolio construct (#15099)
This is the first minimal release of an L2 construct for a service catalog portfolio. The portfolio
construct includes methods to give access to IAM Principals as well as share across accounts. In a future
PR, functionality to associate with a Product and add constraints will be added.

Testing done
------------------
* `yarn build && yarn lint`
* `yarn rosetta:extract --strict`
* `yarn test`
* `yarn integ`

Co-authored-by: Aidan Crank <arcrank@gmail.com>
Co-authored-by: Jackie Wang <wanjacki@amazon.com>


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*