fix(bootstrap): `deploy-role` could be used to directly access buckets in target account
In the previous formulation of the permissions of the `deploy-role`, it
had `s3:GetObject` on `*`, meaning it could read buckets in the target
account.

The permissions were only necessary to read cross-account artifact
buckets created by CDK Pipelines, and so strictly speaking too much
data could be read.

While not strictly an escalation of privileges (given that the user
who is allowed to assume `deploy-role` can already do *anything* to
the account using CloudFormation, and so is in principle already
able to read those buckets via a detour) it is still good practice
to lock the unnecessary direct permissions down.

The new permissions for the `deploy-role` as as follows:

* Can read its own staging bucket (necessary to read the initial
  CloudFormation template for `cdk deploy`)
* Can read S3 buckets *not* in its own account: this enables
  cross-account deployments (by reading the artifact bucket in the
  other account). The nature of cross-account access
  *requires* that the target bucket has a trust policy allowing the
  access as well, so not an exploit.
* Can use all KMS keys as long as it's via S3: makes sure that it
  cannot use KMS keys directly.

In-account bucket accesses are not directly permitted by the principal
roles; however, in-account accesses are enabled by resource policy
statements on the buckets themselves (which are sufficient to extend
the required permissions to the principal).

Fixes #12985, closes #14082.